refactor: change name file App.jsx to router.jsx
