+++ src/main.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import App from '@/routes/App'
+import App from '@/routes/router'
 import '@/styles/index.css'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
